pyproject.toml
[tool.poetry]
name = "trigger-flow"
version = "0.0.1"
description = "Project with FastAPI and Celery"
authors = ["Vladimir faiver90@gmail.com"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.116.1"
redis = "^6.3.0"
httpx = "^0.28.1"
passlib = "^1.7.4"
celery = "^5.5.3"
dotenv = "^0.9.9"
requests = "^2.32.4"
aiosmtplib = "^4.0.1"
pytz = "^2025.2"
ruff = "^0.13.0"
uvicorn = "^0.35.0"
gunicorn = "^23.0.0"
asyncpg = "^0.30.0"
psycopg2-binary = "^2.9.9"
python-jose = { extras = ["pycryptodome"], version = "^3.5.0" }
cryptography = "^46.0.1"
pydantic = { extras = ["email"], version = "^2.11.7" }
python-json-logger = "^3.3.0"
bcrypt = ">=4.0,<4.2"
pydantic-settings = "^2.11.0"
mypy = "^1.18.2"
psycopg = {extras = ["binary"], version = "^3.2.12"}

[tool.poetry.group.dev.dependencies]
alembic = "^1.16.4"
pytest = "^8.4.1"
black = "^25.1.0"
pytest-asyncio = "^1.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
python_files = ["tests.py", "test_*.py", "*_tests.py"]
addopts = """
    --maxfail=9999
    --disable-warnings
    -vv
    -s
    --color=yes
    --continue-on-collection-errors
"""
asyncio_mode = "auto"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501", "B008"]
fixable = ["ALL", "I"]
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.black]
line-length = 88
target-version = ["py311"]


[tool.mypy]
explicit_package_bases = true
ignore_missing_imports = true